backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
ffmpeg-python
google-genai
httpx
pydantic
python-dotenv
